==> 01_importdata.Rmd <==
---
title: "01_importdata"
author: "Michaela Gustafson"
date: "12/30/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## IMPORTING DATA ####

# Description: 
Importing 2019 and 2021 comprehensive prey survey data for the Gyrfalcon and Tundra Conservation Project and the Master's work of Michaela Gustafson.

## LIBRARY

Packages used in this code
```{r library}
library(here)
library(dplyr)
library(tidyr)
```

## IMPORT

Importing environmental and observation data. I used the here() function to call my data saved in my working directory. To use this function, simply download the following datasets into a 'data' folder within your working directory.

```{r import}
obs.main.19 <- read.csv(here("data/Point_Count_Obs_2019_MG_cleaned.csv"))

env.main.19 <- read.csv(here("data/Point_Count_Env_2019_MG_cleaned.csv"))

obs.main.21 <- read.csv(here("data/Point_Count_Obs_2021_MG.csv"))

env.main.21 <- read.csv(here("data/Point_Count_Env_2021_MG.csv"))
```

We can see in the Global Environment that we've ended up with thousands of extra rows.
This was some sort of formatting error in the Excel sheets that I couldn't fix.
We will do an intial clean of these datasets to slim down the tables.


## MINOR CLEANING
Clean up table size by dropping all rows in which the Road_ID column contains NA.
This is only necessary for 2019 environmental and 2019/21 observational.

2019 Environmental Data
```{r 2019envclean}
env.main.19.short <- env.main.19 %>% dplyr::filter(!(Road_ID==""))
```

2019 & 2021 Observation Data
```{r 1921obsclean}
obs.main.19.short <- obs.main.19 %>% 
  dplyr::filter(!(Road_ID==""))

obs.main.21.short <- obs.main.21 %>%
  dplyr::filter(!(ROAD_ID==""))

```


## WRITE NEW DATA CSV

I will write .csv files of these shorter data frames to import into my data cleaning module
and so I will have Excel copies without all the extra blank rows and so I don't actually touch
the raw data

```{r writecsv}
write.csv(env.main.19.short, "data/env_main_19_short.csv")
write.csv(obs.main.19.short, "data/obs_main_19_short.csv")
write.csv(obs.main.21.short, "data/obs_main_21_short.csv")
write.csv(env.main.21, "data/env_main_21_short.csv") # even though there were no excess
# rows in the 2021 environmental data, rewrote a .csv to be consistent with imports for 
# the cleaning module

```


## END SCRIPT
