==> 04_addhabitat.Rmd <==
---
title: "04_addhabitat"
author: "Michaela Gustafson"
date: "1/1/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Adding Habitat Data from ABoVE Dataset

In this code I will be importing GPS points for 2019 and 2021 prey surveys.

I will be adding land cover raster data from: https://daac.ornl.gov/ABOVE/guides/Annual_Landcover_ABoVE.html

1. I will group land cover types into more broad categories.
2. Create a separate layer for each land cover type.
3. Run a focal function for a 390m buffer to be able to extract the proportion of that habitat type within a 390(400m) buffer around each survey point


## LIBRARY
These are the packages used in this code:
```{r library}
library(terra)
library(sf)
library(sp)
library(rgdal)
library(ggplot2)
library(tidyverse)
library(raster)
library(tidyr)
library(dplyr)
library(stringr)
library(rgeos)
```

## IMPORT DEM
Import elevation raster and get land cover and elevation to have same origin, CRS, and extent
```{r dem}
ak_dem <- raster::raster("D:/GYRF/Gyrfalcon/gyrfalcon/data/DEM_Michaela/DEM_Michaela/dsm_mosaic_AK.tif")
```



## UPLOAD LAND COVER RASTERS
I will be selecting out the files that have the land cover for 2014. The four tiles called in below cover the study area.

```{r landcover}

test.tif.14 <- raster::raster("D:/GYRF/Gyrfalcon/gyrfalcon/data/ABoVE_2014lyrs/ABoVE_LandCover_Bh04v01_2014.tif")
test.tif2.14 <- raster::raster("D:/GYRF/Gyrfalcon/gyrfalcon/data/ABoVE_2014lyrs/ABoVE_LandCover_Bh04v02_2014.tif")
test.tif3.14 <- raster::raster("D:/GYRF/Gyrfalcon/gyrfalcon/data/ABoVE_2014lyrs/ABoVE_LandCover_Bh03v01_2014.tif")
test.tif4.14 <- raster::raster("D:/GYRF/Gyrfalcon/gyrfalcon/data/ABoVE_2014lyrs/ABoVE_LandCover_Bh03v02_2014.tif")

r <- raster::merge(test.tif.14, test.tif2.14, test.tif3.14, test.tif4.14)
r[r <= 0 ] <- NA

#plot(r)

writeRaster(r, "D:/GYRF/Gyrfalcon/gyrfalcon/data/ABoVE_2014lyrs/all.mos.tif", overwrite = TRUE)


```


## IMPORT SURVEY LOCATIONS

2019 survey locations are in lat/lon. 2021 survey locations are in UTMs.
I will be converting 2021 points to lat/lon.


### 2019

The recorded lat/lon ended up being in weird locations so I pulled the lat/lon directly from the gps (finalwp_2019.csv) and will be matching the lat/lon waypoints with those recorded in the orignial data sheet ("Point_Count_Env_2019_MG_cleaned.csv")

In 2019 data, there are points not taken for some survey locations so I will be importing the predetermined lat/lon locations for those from the file imported from the GPS (originalpoints_2019.csv)

At the moment, not sure what I used the backup points for...

```{r pointimport}

points.19 <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/Point_Count_Env_2019_MG_cleaned.csv")
gpx.19 <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/finalwp_2019.csv")
ogpoints.19 <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/originalpoints_2019.csv", skip = 22)
backup.19 <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/GPS002_2019backup.csv", skip = 22)

# There are extra '1' at the end of the waypoint number names, need to remove them so the key id will match when merging dataframes
backup.19$name = str_remove(backup.19$name, "^0")


# match key id names
gpx.19 <- rename(gpx.19, Waypoint_num = name)
# add leading zero to key id in points.19 to match key id (Waypoint Name)
points.19$Waypoint_num <- str_pad(points.19$Waypoint_num, 3, pad = "0")
# join points and locations from GPS together
points.gpx.join <- left_join(points.19, gpx.19)

# create id key column
points.gpx.join$id <- paste(points.gpx.join$Road_ID, points.gpx.join$UNIT_ID, points.gpx.join$Transect_ID, points.gpx.join$Point_ID, sep = "_")
head(points.gpx.join)


colnames(points.gpx.join)

# only keep columns need
points.2019 <- points.gpx.join[,c(35, 28:29)]
head(points.2019)
# lat lon are empty because I will be pulling that data in from the correct gpx file or original points file. 

# separating out the first 17 surveyed points that don't have matching waypoint
# names in the gps files
missing.points <- points.gpx.join[c(1:17),]

# separate into KOUG and TELL points for naming purposes
miss.points.koug <- missing.points[c(1:11),]
miss.points.tell <- missing.points[c(12:17),]
# Need id names to match those on the points gpx

# Add K or T to column characters respectively so that columns have matching id keys
# and I can bind/join table together 
miss.points.koug$UNIT_ID <- paste("K", miss.points.koug$UNIT_ID, sep = "")
miss.points.tell$UNIT_ID <- paste("T", miss.points.tell$UNIT_ID, sep = "")
miss.points.koug$Transect_ID <- paste("T", miss.points.koug$Transect_ID, sep = "")
miss.points.tell$Transect_ID <- paste("T", miss.points.tell$Transect_ID, sep = "")


# bind all points back together into one dataframe
missing.points <- rbind(miss.points.koug, miss.points.tell)

# create the universal id column in the dataframe of missing points
missing.points$name <- paste(missing.points$UNIT_ID, missing.points$Transect_ID, missing.points$Point_ID, sep = "_")

# join the missing points to dataframe to data taken off GPS to get true lat/lon
# for these points
miss.og.point.join <- left_join(missing.points, ogpoints.19, by = "name")


# only keep columns needed and rename them
colnames(miss.og.point.join)
miss.og.points <- miss.og.point.join[,c(36, 38,39)]
miss.og.points <- rename(miss.og.points, "lat" = "lat.y")
miss.og.points <- rename(miss.og.points, "lon" = "lon.y")
miss.og.points <- rename(miss.og.points, "id" = "name")


### NEED TO CHANGE Column ID NAMES BACK TO MATCH POINTS 2019

koug.names <- as.data.frame(miss.points.koug$id)
koug.names <- rename(koug.names, 'id' = 'miss.points.koug$id')
tell.names <- as.data.frame(miss.points.tell$id)
tell.names <- rename(tell.names, 'id' = 'miss.points.tell$id')
miss.names <- rbind(koug.names, tell.names)


miss.og.points$id <- miss.names$id
str(miss.og.points)

#######################################################################
# remove those incorrect 'missing points' from the dataframe so I can correctly add them in later
points.2019 <- points.2019[complete.cases(points.2019),]

# add corrected points with true lat/lon
points.2019.all <- rbind(miss.og.points, points.2019)

# save CSV of these fixed points and full df for safe keeping
write.csv(points.2019.all, ("D:/GYRF/Gyrfalcon/gyrfalcon/output/points_2019_all_FINAL.csv"))

## assign crs (pulled from Jenny's version cause I like the way she did it better)

#create an xy dataframe of lat and lon points
xy <- points.2019.all[c(3,2)]
# rename columns 'x' and 'y'
colnames(xy)[1] <- "x"; colnames(xy)[2] <- "y"
# turn into numeric object
xy[,1:2] <- sapply(xy[,1:2], as.numeric)
#turn into coordinates string
coordinates(xy) <- ~ x + y
# now turn into spatial points dataframe matching the newly made saptial points data frame coordinates to the associated data from the original dataframe
spdf <- SpatialPointsDataFrame(coords = xy, data = points.2019.all)
# assing CRS that points were collected in on GPS
proj4string(spdf) <- CRS("+init=epsg:4326")
# reprojeting it to DEM crs first to create extent to crop DEM to more manageable size
points19.proj <- spTransform(spdf, crs(ak_dem)) 
plot(ak_dem); points(points19.proj)
writeOGR(points19.proj, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts19" , driver="ESRI Shapefile")

# now reproject to crs of land cover data and save
points19.proj.lc <- spTransform(spdf, crs(r))
writeOGR(points19.proj.lc, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts19.proj.lc" , driver="ESRI Shapefile")


```


### 2021

```{r 21points}

##### 2021 #####
points.21 <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/Point_Count_Env_2021_MG.csv")
head(points.21); dim(points.21)

# subset id and utms
colnames(points.21)
points.21 <- points.21[,c(1:4, 19:20)]
points.21$id <- paste(points.21$ROAD_ID, points.21$UNIT_ID, points.21$TRANSECT_ID, points.21$POINT_ID, sep = "_")

head(points.21)
#subset again to keep only final key id and utms
points.21 <- points.21[,c(7, 5:6)]
head(points.21); dim(points.21)
# export for safe keeping
write.csv(points.21, "D:/GYRF/Gyrfalcon/gyrfalcon/output/utms_21_final.csv")

# now do same thing as with 2019 data to assign CRS
xy2 <- points.21[c(2,3)]
colnames(xy2)[1] <- "x"; colnames(xy2)[2] <- "y"
coordinates(xy2) <- ~ x + y
spdf2 <- SpatialPointsDataFrame(coords = xy2, data = points.21)
proj4string(spdf2) <- CRS("+init=epsg:26903")
points21.proj <- spTransform(spdf2, crs(ak_dem))
plot(ak_dem); points(points21.proj)
writeOGR(points21.proj, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts21" , driver="ESRI Shapefile")
points21.proj.lc <- spTransform(spdf2, crs(r))
writeOGR(points21.proj.lc, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts21.proj.lc" , driver="ESRI Shapefile")


```


### Bind 2019 and 2021 points together

```{r ptbind}
points19.proj.lc@data$year <- 2019
points21.proj.lc@data$year <- 2021
new <- bind(points19.proj.lc[c(1,4)], points21.proj.lc[c(1,4)])
writeOGR(new, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts19&21.proj.lc" , driver="ESRI Shapefile")

```

### Transform CRS of points to match DEM
```{r pointstodem}
# transform crs to match dem
pts.all <- readOGR("D:/GYRF/Gyrfalcon/gyrfalcon/data", "pts19&21.proj.lc")
pts.all2 <- spTransform(pts.all, CRS("+proj=aea +lat_0=50 +lon_0=-154 +lat_1=55 +lat_2=65 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs"))
writeOGR(pts.all2, dsn="D:/GYRF/Gyrfalcon/gyrfalcon/data", layer="pts19&21" , driver="ESRI Shapefile")

```

### Aggregate to 400m resolution for DEM and land cover rasters

```{r resample400}
####### resample/aggregate dem to 400 m resolution using mean value ########
r <- raster(extent(ak_dem), res = 400, crs = "+proj=aea +lat_0=50 +lon_0=-154 +lat_1=55 +lat_2=65 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs ")

dem.400 <- aggregate(ak_dem, fact = 80, fun = mean)

writeRaster(dem.400, "D:/GYRF/Gyrfalcon/gyrfalcon/data/dem.400.tif", overwrite = TRUE)

############### create land cover percent layers 400 meter resolution ########
all.mos <- raster("D:/GYRF/Gyrfalcon/gyrfalcon/data/ABoVE_2014lyrs/all.mos.tif")

cov_pct <- lapply(unique(all.mos), function(land_class) {
  aggregate(all.mos, fact=13, fun=function(values, na.rm) {
    sum(values==land_class, na.rm=na.rm)/length(values)
  })
})

cov_pct2 <- list()
for(i in 1:length(cov_pct)){
  cov_pct2[[i]] <- projectRaster(cov_pct[[i]], crs = "+proj=aea +lat_0=50 +lon_0=-154 +lat_1=55 +lat_2=65 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs ")
}

lc <- list()
for(i in 1:length(cov_pct2)){
  lc[[i]] <- resample(cov_pct2[[i]], r, method = "bilinear")
}

names(lc) <- c("egforest", "decforest", "mixforest", "woodland", "lowshrub", "tallshrub", "openshrub", "tundra", "tussock", "sparseveg", "fen", "bog", "littoral", "barren", "water")

lc <- stack(lc)

### resample so all rasters line up ######
dem.400.re <- resample(dem.400, lc[[1]], method = "bilinear")

c <- stack(lc, dem.400.re)

# save each layer as it own tif
writeRaster(c, filename=names(c), bylayer=TRUE, format="GTiff", overwrite = TRUE)

# will copy these tifs over to GYRF/Gyrfalcon/gyrfalon folder manually...
```

### Final part - extracting at points
```{r extract}
########### bring in all resampled 400 m rasters to extract to survey points #########
######### extract by 400 m buffer around all points ######
rastlist <- list.files(path = "D:/GYRF/Gyrfalcon/gyrfalcon/400m", pattern='.tif$', all.files=TRUE, full.names = TRUE)
all <- lapply(rastlist, raster)
a <- stack(all)

pts <- readOGR("D:/GYRF/Gyrfalcon/gyrfalcon/data", "pts19&21")

pts.extract <- as.data.frame(raster::extract(a, pts, buffer = 400, fun=mean))
df <- as.data.frame(pts.all2)
pts.covs <- cbind(df, pts.extract)

colnames(pts.covs)[8] <- "elev"

write.csv(pts.covs, "D:/GYRF/Gyrfalcon/gyrfalcon/output/pts.covs.csv")
```
