==> 07_finalenvjoin.Rmd <==
---
title: "07_finalenvjoin"
author: "Michaela Gustafson"
date: "1/4/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# FINAL JOIN OF ENVIRONMENTAL VARIABLES

In this code I will be joining my extracted land cover and elevation values to my full environmental dataset to be used in my time removal models.


1. Import full environmental dataframe
2. Import 2019&2021 land cover and elevation covariates
4. Merge and export dataframe


## LIBRARY
These are the following packages used in this code:
```{r library}

library(dplyr) # for manipulating data
```


## 1. Import environmental dataframe
Importing from previous cleaning code "data/all_env.csv"
```{r envimport}
all.env <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/all_env.csv")
# take out added 'X' column
colnames(all.env)
all.env <- all.env[,c(2:10)]
# check
head(all.env)

### Putting in another cleaning section here so I don't have to completely rerun like 6 other painstakingly long codes

##### NEED TO CHANGE SKY VARIABLE TO HIGHEST NUMBER

str(all.env)
all.env$sky <- as.factor(all.env$sky)


# going to have to use dsub and do it for each character string of multiple categories
unique(all.env$sky)

all.env$sky <- gsub("2,3,4", "4", all.env$sky)
all.env$sky <- gsub("3,4", "4", all.env$sky)
all.env$sky <- gsub("2,4", "4", all.env$sky)
all.env$sky <- gsub("2,3", "3", all.env$sky)
all.env$sky <- gsub("0,3", "3", all.env$sky)
all.env$sky <- gsub("1,3", "3", all.env$sky)
all.env$sky <- gsub("3,0", "3", all.env$sky)
all.env$sky <- gsub("3,", "3", all.env$sky)

unique(all.env$sky)

all.env$sky <- as.factor(all.env$sky)
```

### 2. Import land cover

```{r lc}
pt.covs <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/output/pts.covs.csv")

# get rid of extra 'X' column from import
colnames(pt.covs)
pt.covs <- pt.covs[,c(2:21)]

# check
head(pt.covs); dim(pt.covs)

# I know there is a spelling error for COUN_3_12_4 row 38, col 1 in all.lc dataframe
pt.covs[38,1] <- "COUN_3_12_4"

```


### 4. Merge land cover and elevation to all environmental

```{r allmerge}
all.covs <- left_join(all.env, pt.covs, by = 'id')

# and from future  coding I know there is a missing temperature
# double check that its in the same location and input the correct 
# value from the hard copy datasheets
x <- as.data.frame(is.na(all.covs))
which(x == "TRUE", arr.ind = TRUE)
# one of my temps is NA
# looked at temps of surveys before (40) and after (44) and can enter a temp
all.covs[552,6] <- "42"

```


3. Merge (add) land cover types into specified groups
```{r}
colnames(all.covs)
all.covs2 <- all.covs %>%
  mutate(forest = egforest + decforest + mixforest + woodland,
         bare = barren + sparseveg,
         wetland = bog + fen + littoral)
# take out extra columns (ones that were added together) and reorder
colnames(all.covs2)
all.covs3 <- all.covs2[,c(1:9, 11, 12, 16, 20, 22, 24:27, 29:31)]


### export final dataframe:

write.csv(all.covs3, "D:/GYRF/Gyrfalcon/gyrfalcon/output/all.covs.FINAL.csv")
```


# END CODE
