==> 08_squir.Rmd <==
---
title: "08_squir"
author: "Michaela Gustafson"
date: "1/4/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## TIME REMOVAL ANALYSIS: ARCTIC GROUND SQUIRREL

In this code I will be running my time-removal abundance model for Arctic Ground Squirrels.



## LIBRARY
These are the packages used in the following code:
```{r library}
library(dplyr)
library(tidyverse)
library(ggplot2)
library(unmarked)
library(AICcmodavg)
```


## LOAD OBSERVATION and ENVIRONMENTAL DATA

```{r loaddata}
all.obs <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/data/all_obs.csv")
all.env <- read.csv("D:/GYRF/Gyrfalcon/gyrfalcon/output/all.covs.FINAL.csv")
# remove extra 'X' column from import
all.obs <- all.obs[,-c(1)]
all.env <- all.env[,-c(1)]
# inspect

head(all.obs)
head(all.env)

# looks good
```

## PREP OBSERVATION DATA FOR SELECTED SPECIES

```{r squirprep}
# copy data table for SQUIR

squir <- data.frame(all.obs)

# change all species codes of SQUIR to a 1 and all other species to 0
squir$species[squir$species == "SQUIR"] <- 1
squir$species[squir$species != "1"] <- 0

# change count to 0 for non SQUIR
squir$count[squir$species == 0] <- 0
head(squir)
# now 'squir' df represents counts for SQUIR

# check for na values
sum(is.na(squir$time_int))


# change time intervals to reflect intervals of 2 minutes
squir2 <- mutate(squir, 
                time_int = ifelse(squir$time_int %in% 0:1, "1", 
                                  ifelse(squir$time_int %in% 2:3, "2", 
                                         ifelse(squir$time_int %in% 4:5, "3",
                                                ifelse(squir$time_int %in% 6:7, "4",
                                                       ifelse(squir$time_int %in% 8:9, "5", "NA"))))))


# aggregate rows and sum counts for matching keys and time intervals
# must change formats first:

str(squir2)

squir2$id <- as.factor(squir2$id)
squir2$time_int <- as.factor(squir2$time_int)
squir2$count <- as.integer(squir2$count)

squir.agg <- aggregate(x = squir2$count, 
                      by = list(squir2$id, squir2$time_int), 
                      FUN = sum)




# rename columns in aggregated df

head(squir.agg)

names(squir.agg)[names(squir.agg) == "Group.1"] <- "id"
names(squir.agg)[names(squir.agg) == "Group.2"] <- "time_int"
names(squir.agg)[names(squir.agg) == "x"] <- "count"

head(squir.agg)

# okay so squir.agg is our count dataframe, we don't need any of the other columns,
# those were only used to help filter out for distance and flyover observatiosn

# check that ids are matching between env and obs
# need to change id to a factor to use anti-join
all.env$id <- as.factor(all.env$id)
miss1 <- anti_join(all.env, squir.agg) # clear
miss2 <- anti_join(squir.agg, all.env) # clear

# nothing missing! 


# spread dataframes:

unique(squir.agg$id) # should end up with 916 rows


squir.wide <- squir.agg %>%
  dplyr::select(id, time_int, count) %>%
  spread(key = time_int, value = count, fill = 0)



```

### PREP ENVIRONMENTAL DATA
```{r prepenv}
# check structure of variables

str(all.env) 
all.env$sky <- as.factor(all.env$sky)
all.env$observer <- as.factor(all.env$observer)
all.env$hear <- as.factor(all.env$hear)

```

## CHECK DATA FOR CORRELATIONS, ETC.

### Look at land cover types

```{r corrtest}

# scale numerical predictors
#scale predictors:
str(all.env)

jul.scaled <- scale(all.env$julian)
min.scaled <- scale(all.env$min_after_sun)
temp.scaled <- scale(as.numeric(all.env$tempf))
wind.scaled <- scale(all.env$wind)
t.scaled <- scale(all.env$tundra)
tuss.scaled <- scale(all.env$tussock)
ls.scaled <- scale(all.env$lowshrub)
ts.scaled <- scale(all.env$tallshrub)
os.scaled <- scale(all.env$openshrub)
b.scaled <- scale(all.env$bare)
f.scaled <- scale(all.env$forest)
wl.scaled <- scale(all.env$wetland)
wat.scaled <- scale(all.env$water)
elev.scaled <- scale(all.env$elev)

# create scaled dataframe
all.env.scaled <- data.frame(all.env)

# replace variables with scaled ones
all.env.scaled$julian <- jul.scaled
all.env.scaled$min_after_sun <- min.scaled
all.env.scaled$tempf <- temp.scaled
all.env.scaled$wind <- wind.scaled
all.env.scaled$tundra <- t.scaled
all.env.scaled$tussock <- tuss.scaled
all.env.scaled$lowshrub <- ls.scaled
all.env.scaled$tallshrub <- ts.scaled
all.env.scaled$openshrub <- os.scaled
all.env.scaled$bare <- b.scaled
all.env.scaled$forest <- f.scaled
all.env.scaled$wetland <- wl.scaled
all.env.scaled$water <- wat.scaled
all.env.scaled$elev <- elev.scaled

str(all.env.scaled)
# correlation matrix
cor.matrix <- as.matrix(all.env.scaled[,c(2, 4, 6, 8, 10:19)])
cor.mat <- cor(cor.matrix) 
cor.mat
```

## TIME-REMOVAL MODEL

```{r trmodel}

# check structures and join data frames
colnames(all.env.scaled)
str(all.env.scaled)
str(squir.wide)

squir.env.join <- left_join(squir.wide, all.env.scaled)
head(squir.env.join); dim(squir.env.join)
str(squir.env.join)

# create observation and site covariate data frames
colnames(squir.env.join)

timeints <- squir.env.join[,c(2:6)]
siCovs <- squir.env.join[,c(7, 9:24)]


squirFrame2 <- unmarkedFrameMPois(
  # import time removal columns(counts):
  y = timeints, 
  siteCovs = siCovs, 
  # define pifun type: 
  type = "removal" )

# fit models: multinomPois order of formulas: detection, abundance
# I will only be running a null model and a full model. Not gonna mess around with that multiple model bullshit
# can't use all land cover variables because htey add up to 1 within each site

fm0 <- multinomPois(~ 1 ~ 1, data = squirFrame2) #null model
fm.full <- multinomPois( ~ -1 + julian + min_after_sun + observer + sky + wind + hear ~ -1 + tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev, data = squirFrame2)
# LOOK UP WARNING MESSAGE

ms2 <- fitList(
  "lam(.)p(.)" = fm0,
  "lam(tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev)p(julian + min_after_sun + observer + sky + wind + hear)" = fm.full)

(ms2sel <- modSel(ms2))

summary(fm.full)
plot(fm.full)
```

## CHECK MODEL FIT
Will be looking at c-hat for overdispersion
```{r gfone, eval=FALSE}
# Goodness of fit: C-hat result 1.1

chat <- Nmix.gof.test(fm.full, nsim = 1000, plot.hist = TRUE, parallel = TRUE, ncores = 5)
chat # 2.2

```

## NEGATIVE BINOMIAL MODEL

Running a negative binomial model through the 'gmultimix' function. 

First, must set up a GMM dataframe

```{r negbinom, eval=FALSE}
# set up dataframe

NBsquirFrame <- unmarkedFrameGMM(
  y = timeints, 
  siteCovs = siCovs, 
  # now have this number of primary sampling periods argument we must fill:
  numPrimary=1,
  # define pifun type: 
  type = "removal" )

# now fit negative binomial models:
fm0nb <- gmultmix(~1, ~1, ~1, mixture = "NB", data = NBsquirFrame)
fm.full.nb <- gmultmix( ~ -1 + tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev, ~ 1, ~ -1 + julian + min_after_sun + observer + sky + wind + hear,mixture = "NB",data = NBsquirFrame)


# rank by AIC
gms <- fitList(
  "NB,lam(.)p(.)" = fm0nb,
  "NB,lam(tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev)p(julian + min_after_sun + observer + sky + wind + hear)" = fm.full.nb
)

gms1 <- modSel(gms)
output.nb <- as(gms1, "data.frame")

output.nb


summary(fm.full.nb)
fm.full.nb

# check c-hat for overdispersion?

chat.nb <- Nmix.gof.test(fm.full.nb, nsim = 1000, plot.hist = TRUE, parallel = TRUE, ncores = 5)
chat.nb #0.85
```
 
### ZERO INFLATED POISSON
For this I will need to fit the data into a different unmarked Frame

```{r zipframe}

squirFrameZIP <- unmarkedFramePCount(
  y = timeints,
  siteCovs = siCovs
)
squirFrameZIP
summary(squirFrameZIP)

```

Now fit the model using pcount()
```{r zipmodel}
fm0ZIP <- pcount(~1~1, squirFrameZIP, mixture = "ZIP")
fmZIP <- pcount(~ -1 + julian + min_after_sun + observer + sky + wind + hear, ~ -1 + tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev, squirFrameZIP)
```
