==> 10_ropt.Rmd <==
---
title: "10_ropt"
author: "Michaela Gustafson"
date: "1/9/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## TIME REMOVAL ANALYSIS: WILLOW PTARMIGAN

In this code I will be running my time-removal abundance model for Willow Ptarmigan.

This is a multinomial N-mixture model.

## LIBRARY
These are the packages used in the following code:
```{r library}
library(here) # for locating filepath
library(dplyr)
library(tidyverse)
library(ggplot2)
library(unmarked)
library(AICcmodavg)
```


## LOAD OBSERVATION and ENVIRONMENTAL DATA

```{r loaddata}
all.obs <- read.csv(here("data/all_obs.csv"))
all.env <- read.csv(here("data/all_env_FINAL.csv"))

# inspect

head(all.obs)
head(all.env)

# get rid of extra 'X' column added from import

all.obs <- all.obs[,c(2:10)]
all.env <- all.env[,c(2:20)]

# check
head(all.obs)
head(all.env)

# looks good
```
## PREP OBSERVATION DATA FOR SELECTED SPECIES

```{r roptprep}
# copy data table for Rock Ptarmigan

ropt <- data.frame(all.obs)

# change all species codes of ROPT to a 1 and all other species to 0
ropt$species[ropt$species == "ROPT"] <- 1
ropt$species[ropt$species != "1"] <- 0

# change count to 0 for non WIPT
ropt$count[ropt$species == 0] <- 0
head(ropt)

sum(ropt$count)
# sample size of 62
# now 'ropt' df represents counts for Rock Ptarmigan

# check for na values
sum(is.na(ropt$time_int))


# change time intervals to reflect intervals of 2 minutes
ropt2 <- mutate(ropt, 
                time_int = ifelse(ropt$time_int %in% 0:1, "1", 
                                  ifelse(ropt$time_int %in% 2:3, "2", 
                                         ifelse(ropt$time_int %in% 4:5, "3",
                                                ifelse(ropt$time_int %in% 6:7, "4",
                                                       ifelse(ropt$time_int %in% 8:9, "5", "NA"))))))


# aggregate rows and sum counts for matching keys and time intervals
# must change formats first:

str(ropt2)

ropt2$id <- as.factor(ropt2$id)
ropt2$time_int <- as.factor(ropt2$time_int)
ropt2$count <- as.integer(ropt2$count)

ropt.agg <- aggregate(x = ropt2$count, 
                      by = list(ropt2$id, ropt2$time_int), 
                      FUN = sum)




# rename columns in aggregated df

head(ropt.agg)

names(ropt.agg)[names(ropt.agg) == "Group.1"] <- "id"
names(ropt.agg)[names(ropt.agg) == "Group.2"] <- "time_int"
names(ropt.agg)[names(ropt.agg) == "x"] <- "count"

head(ropt.agg)

# okay so ropt.agg is our count dataframe, we don't need any of the other columns,
# those were only used to help filter out for distance and flyover observatiosn

# check that ids are matching between env and obs
# need to change id to a factor to use anti-join
all.env$id <- as.factor(all.env$id)
miss1 <- anti_join(all.env, ropt.agg) # clear
miss2 <- anti_join(ropt.agg, all.env) # clear



# spread dataframes:

unique(ropt.agg$id) # should end up with 916 rows


ropt.wide <- ropt.agg %>%
  dplyr::select(id, time_int, count) %>%
  spread(key = time_int, value = count, fill = 0)



```

### PREP ENVIRONMENTAL DATA
```{r prepenv}
# check structure of variables

str(all.env) 
all.env$sky <- as.factor(all.env$sky)
all.env$year <- as.factor(all.env$year)
all.env$observer <- as.factor(all.env$observer)
all.env$hear <- as.factor(all.env$hear)

```


## CHECK DATA FOR CORRELATIONS

Do not need to repeat this part since environmental data is the exact same as in the first code run for ground squirrels

### BUT DO NEED TO SCALE PREDICTORS
I will export the scaled environmental dataframe for use in future codes

```{r scaledf}
# bring in scaled dataframe saved in 09_wipt

all.env.scaled <- read.csv(here("data/all_env_scaled_df.csv"))
```

## TIME-REMOVAL MODEL

ROPT sample size: 62
Might need to run a zero inflated poisson on this model as well.
```{r trmodel}

# check structures and join data frames
colnames(all.env.scaled)
str(all.env.scaled)
str(ropt.wide)

ropt.env.join <- left_join(ropt.wide, all.env.scaled)
head(ropt.env.join); dim(ropt.env.join)
str(ropt.env.join)

# create observation and site covariate data frames
colnames(ropt.env.join)

timeints <- ropt.env.join[,c(2:6)]
siCovs <- ropt.env.join[,c(8, 10:25)]


roptFrame2 <- unmarkedFrameMPois(
  # import time removal columns(counts):
  y = timeints, 
  siteCovs = siCovs, 
  # define pifun type: 
  type = "removal" )

summary(roptFrame2)
# sample size: 62

# fit models: multinomPois order of formulas: detection, abundance
# I will only be running a null model and a full model. Not gonna mess around with that multiple model bullshit
# can't use all land cover variables because htey add up to 1 within each site

fm0 <- multinomPois(~ 1 ~ 1, data = roptFrame2) #null model
fm.full <- multinomPois( ~ -1 + julian + min_after_sun + observer + sky + wind + hear ~ -1 + tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev, data = roptFrame2)
# LOOK UP WARNING MESSAGE

ms2 <- fitList(
  "lam(.)p(.)" = fm0,
  "lam(tundra + tussock + lowshrub + tallshrub + openshrub + bare + wetland + elev)p(julian + min_after_sun + observer + sky + wind + hear)" = fm.full)

(ms2sel <- modSel(ms2))

summary(fm.full)

```

## CHECK MODEL FIT
Will be looking at c-hat for overdispersion, as well as chisq value and its p-value
```{r gfone, eval=FALSE}
# Goodness of fit

chat <- Nmix.gof.test(fm.full, nsim = 1000, plot.hist = TRUE, parallel = TRUE, ncores = 7)
chat # 0.97

```
